bootloader: Refactor into multiple C files
